+++ vcd/provider.go
@@ -119,6 +119,7 @@ func Provider() terraform.ResourceProvider {
 			"vcd_lb_virtual_server":  resourceVcdLBVirtualServer(),  // 2.4
 			"vcd_nsxv_dnat":          resourceVcdNsxvDnat(),         // 2.5
 			"vcd_nsxv_snat":          resourceVcdNsxvSnat(),         // 2.5
+			"vcd_nsxv_firewall_rule": resourceVcdNsxvFirewallRule(), // 2.5
 		},
 
 		DataSourcesMap: map[string]*schema.Resource{
@@ -141,6 +142,7 @@ func Provider() terraform.ResourceProvider {
 			"vcd_lb_virtual_server":  datasourceVcdLbVirtualServer(),  // 2.4
 			"vcd_nsxv_dnat":          datasourceVcdNsxvDnat(),         // 2.5
 			"vcd_nsxv_snat":          datasourceVcdNsxvSnat(),         // 2.5
+			"vcd_nsxv_firewall_rule": datasourceVcdNsxvFirewallRule(), // 2.5
 		},
 
 		ConfigureFunc: providerConfigure,